created utilities.h for shared common functions
