+++ JoySender++/JoySender++.h
@@ -31,6 +31,7 @@ THE SOFTWARE.
 #include "TCPConnection.h"
 #include "ArgumentParser.hpp"
 #include "FPSCounter.hpp"
+#include "utilities.hpp"
 
 #pragma comment(lib, "SDL3.lib")
 
@@ -75,73 +76,11 @@ bool updateRumble(const char motor, BYTE val);
 // Main Loop Function 
 int joySender(Arguments& args);
 
-// Function safely return environment variables
-std::string g_getenv(const char* variableName) {
-    std::string value;
-    size_t requiredSize = 0;
-    getenv_s(&requiredSize, nullptr, 0, variableName);
-    if (requiredSize > 0) {
-        value.resize(requiredSize);
-        getenv_s(&requiredSize, &value[0], requiredSize, variableName);
-        value.pop_back(); // Remove the null-terminating character
-    }
-    return value;
-}
-// Some Console Output, ANSI helper functions
-void enableANSI()
-{
-    HANDLE consoleHandle = GetStdHandle(STD_OUTPUT_HANDLE);
 
-    DWORD consoleMode;
-    GetConsoleMode(consoleHandle, &consoleMode);
-
-    // Enable ANSI escape sequences support
-    consoleMode |= ENABLE_VIRTUAL_TERMINAL_PROCESSING;
-
-    SetConsoleMode(consoleHandle, consoleMode);
-}
-void hideConsoleCursor() {
-    HANDLE consoleHandle = GetStdHandle(STD_OUTPUT_HANDLE);
-    CONSOLE_CURSOR_INFO cursorInfo;
-    GetConsoleCursorInfo(consoleHandle, &cursorInfo);
-    cursorInfo.bVisible = FALSE;
-    SetConsoleCursorInfo(consoleHandle, &cursorInfo);
-}
-void showConsoleCursor() {
-    HANDLE consoleHandle = GetStdHandle(STD_OUTPUT_HANDLE);
-    CONSOLE_CURSOR_INFO cursorInfo;
-    GetConsoleCursorInfo(consoleHandle, &cursorInfo);
-    cursorInfo.bVisible = TRUE;
-    SetConsoleCursorInfo(consoleHandle, &cursorInfo);
-}
-void clearConsoleScreen()
-{
-    std::cout << "\033[2J\033[H";
-}
-void repositionConsoleCursor(int lines, int offset) {
-    if (lines < 0) // Move the cursor up by the specified number of lines
-        std::cout << "\033[" << abs(lines) << "F";
-    else if (lines > 0) // Move the cursor down by the specified number of lines
-        std::cout << "\033[" << lines << "E";
-    // Move the cursor to the specified offset from the start of the line
-    std::cout << "\033[" << offset << "G";
-}
-void clearConsoleLine() {
-    std::cout << "\033[K";
-}
 // Function to display g_outputText to console
 void displayOutputText() {
     clearConsoleScreen();
     std::cout << g_outputText;
-}
-// Function to detect if keyboard key is pressed
-bool getKeyState(int KEYCODE) {
-    return GetAsyncKeyState(KEYCODE) & 0x8000;
-}
-// Function to convert strings to wide strings
-std::wstring g_toWide(std::string& str) {
-    std::wstring wideString(str.begin(), str.end());
-    return wideString;
 }
 // Function that will prompt for and verify an ip address 
 std::string getHostAddress() {
@@ -172,60 +111,6 @@ std::string getHostAddress() {
     return host_address;
 }
 // Function that takes string representing a float and fixes it's decimal places to numDigits
-std::string formatDecimalString(const std::string& str, UINT8 numDigits) {
-    size_t decimalPos = str.find('.');
-    if (decimalPos == std::string::npos) {
-        // No decimal point found, return the original 
-        // string with decimal and padded zeros
-        std::string paddedStr = str + ".";
-        paddedStr.append(numDigits, '0');
-        return paddedStr;
-    }
-    size_t strLen = str.length();
-    size_t numDecimals = strLen - decimalPos - 1;
-    if (numDecimals <= numDigits) {
-        // No need to truncate or pad, return the original string
-        return str;
-    }
-    // Truncate the string to the desired number of decimal places
-    std::string truncatedStr = str.substr(0, decimalPos + numDigits + 1);
-    // Pad with zeros if necessary
-    if (numDecimals < numDigits) {
-        truncatedStr.append(numDigits - numDecimals, '0');
-    }
-
-    return truncatedStr;
-}
-// Function captures all input from the keyboard to clear buffer
-void swallowInput()
-{
-    while (_kbhit())
-    {
-        _getch();
-    }
-}
-// Function waits for no keyboard presses to be detected before returning
-void wait_for_no_keyboard_input() {
-    bool input = true;
-    while (input) {
-        input = false;
-        // Check if any key is currently being pressed
-        for (int i = 48; i < 90; i++) {
-            if (GetAsyncKeyState(i) & 0x8000) {
-                input = true;
-            }
-        }
-    }
-    swallowInput();
-}
-// Function determines if app is the active window
-bool IsAppActiveWindow()
-{
-    HWND consoleWindow = GetConsoleWindow();
-    HWND foregroundWindow = GetForegroundWindow();
-
-    return (consoleWindow == foregroundWindow);
-}
 // Function determines if rumble values need to be updated
 bool updateRumble(const char motor, BYTE val) {
     static BYTE L, R = 0;
